cleaned up code somewhat
